# backend/app/services/migration_advisor/migration_advisor/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Cloud Migration Advisor API
  description: |
    Comprehensive API for the Cloud Migration Advisor platform that guides organizations through
    their cloud migration journey from initial assessment through provider selection, migration
    execution, and post-migration resource organization.
    
    ## Features
    - Migration project management and assessment
    - Workload and requirements analysis
    - ML-powered cloud provider recommendations
    - Migration planning and execution tracking
    - Automated resource discovery and organization
    - Multi-dimensional resource management
    - Post-migration FinOps integration
    
    ## Authentication
    All endpoints require authentication using Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
  version: 1.0.0
  contact:
    name: Cloud Migration Advisor Support
    email: support@cloudmigration.example.com
  license:
    name: Proprietary
servers:
  - url: https://api.cloudmigration.example.com/v1
    description: Production server
  - url: https://staging-api.cloudmigration.example.com/v1
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail
...
    OrganizationProfile:
      type: object
      properties:
        company_size:
          type: string
          enum: [small, medium, large, enterprise]
          description: Company size category
        industry:
          type: string
          description: Industry sector
        current_infrastructure:
          type: string
          enum: [on_premises, cloud, hybrid, multi_cloud]
          description: Current infrastructure type
        it_team_size:
          type: integer
          minimum: 0
          description: Number of IT team members
        cloud_experience_level:
          type: string
          enum: [none, beginner, intermediate, advanced]
          description: Team's cloud experience level
        geographic_presence:
          type: array
          items:
            type: string
          description: List of regions/countries
        additional_context:
          type: object
          additionalProperties: true
          description: Additional context information
      required:
        - company_size
        - industry
        - current_infrastructure
        - it_team_size
        - cloud_experience_level
...
    CategorizedResource:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Resource UUID
        resource_id:
          type: string
          description: Cloud provider resource ID
        resource_type:
          type: string
          description: Type of resource (ec2, s3, rds, etc.)
        resource_name:
          type: string
          description: Resource name
        provider:
          type: string
          description: Cloud provider
        team:
          type: string
          description: Team assignment
        project:
          type: string
          description: Project assignment
        environment:
          type: string
          description: Environment (dev, staging, prod)
        region:
          type: string
          description: Cloud region
        cost_center:
          type: string
          description: Cost center assignment
        custom_attributes:
          type: object
          additionalProperties: true
          description: Custom attributes
        tags:
          type: object
          additionalProperties:
            type: string
          description: Resource tags
        ownership_status:
          type: string
          enum: [assigned, unassigned, suggested, pending]
          description: Ownership status
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - resource_id
        - resource_type
        - provider
        - ownership_status
security:
  - BearerAuth: []
paths:
...
  /api/migrations/{project_id}/assessment/organization:
    post:
      tags:
        - Migration Assessment
      summary: Create organization profile
      description: |
        Create organization profile and estimate assessment timeline.
        Collects comprehensive organization information.
        
        **Requirements:** 1.2, 1.3, 1.5
      operationId: createOrganizationProfile
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationProfile'
            example:
              company_size: "medium"
              industry: "Financial Services"
              current_infrastructure: "on_premises"
              it_team_size: 15
              cloud_experience_level: "beginner"
              geographic_presence: ["North America", "Europe"]
              additional_context:
                compliance_requirements: ["SOC2", "GDPR"]
      responses:
        '200':
          description: Organization profile created with timeline estimation
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    $ref: '#/components/schemas/OrganizationProfile'
                  timeline_estimation:
                    type: object
                    properties:
                      estimated_days:
                        type: integer
                      estimated_completion_date:
                        type: string
                        format: date-time
                      breakdown:
                        type: object
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
          
    get:
      tags:
        - Migration Assessment
      summary: Get organization profile
      description: |
        Retrieve organization profile for a migration project.
        
        **Requirements:** 1.2
      operationId: getOrganizationProfile
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
      responses:
        '200':
          description: Organization profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationProfile'
        '404':
          description: Profile not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...
  /api/migrations/{project_id}/resources/discover:
    post:
      tags:
        - Resource Organization
      summary: Discover cloud resources
      description: |
        Discover and inventory all cloud resources from the specified provider.
        
        **Requirements:** 5.1
      operationId: discoverResources
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum: [aws, gcp, azure]
                  description: Cloud provider
                credentials:
                  type: object
                  additionalProperties:
                    type: string
                  description: Provider credentials
                regions:
                  type: array
                  items:
                    type: string
                  description: Regions to scan
                resource_types:
                  type: array
                  items:
                    type: string
                  description: Resource types to discover
              required:
                - provider
                - credentials
            example:
              provider: "aws"
              credentials:
                access_key_id: "AKIAIOSFODNN7EXAMPLE"
                secret_access_key: "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
              regions: ["us-east-1", "us-west-2"]
              resource_types: ["ec2", "s3", "rds"]
      responses:
        '200':
          description: Resources discovered
          content:
            application/json:
              schema:
                type: object
                properties:
                  discovery_id:
                    type: string
                  provider:
                    type: string
                  total_discovered:
                    type: integer
                  resources_by_type:
                    type: object
                    additionalProperties:
                      type: integer
                  regions_scanned:
                    type: array
                    items:
                      type: string
                  discovery_timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid request
        '404':
          description: Project not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/migrations/{project_id}/resources:
    get:
      tags:
        - Resource Organization
      summary: Get resource inventory
      description: |
        Get resource inventory with optional filtering.
        
        **Requirements:** 5.1, 5.2
      operationId: getResourceInventory
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
        - name: resource_type
          in: query
          description: Filter by resource type
          schema:
            type: string
        - name: ownership_status
          in: query
          description: Filter by ownership status
          schema:
            type: string
            enum: [assigned, unassigned, suggested, pending]
        - name: team
          in: query
          description: Filter by team
          schema:
            type: string
        - name: project_filter
          in: query
          description: Filter by project
          schema:
            type: string
        - name: environment
          in: query
          description: Filter by environment
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Resource inventory
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_resources:
                    type: integer
                  resources_by_type:
                    type: object
                    additionalProperties:
                      type: integer
                  resources_by_status:
                    type: object
                    additionalProperties:
                      type: integer
                  resources:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategorizedResource'
        '404':
          description: Project not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...